server: move pocketIO into internal
