+++ server/BUILD.bazel
@@ -149,7 +149,6 @@ go_test(
         "main_test.go",
         "mock_conn_test.go",
         "optimize_trace_test.go",
-        "packetio_test.go",
         "plan_replayer_test.go",
         "server_test.go",
         "stat_test.go",
@@ -228,7 +227,6 @@ go_test(
         "@com_github_docker_go_units//:go-units",
         "@com_github_go_sql_driver_mysql//:mysql",
         "@com_github_gorilla_mux//:mux",
-        "@com_github_klauspost_compress//zstd",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_kvproto//pkg/kvrpcpb",
