+++ server/internal/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "internal",
@@ -14,5 +14,20 @@ go_library(
         "//sessionctx/variable",
         "@com_github_klauspost_compress//zstd",
         "@com_github_pingcap_errors//:errors",
+    ],
+)
+
+go_test(
+    name = "internal_test",
+    timeout = "short",
+    srcs = ["packetio_test.go"],
+    embed = [":internal"],
+    flaky = True,
+    shard_count = 4,
+    deps = [
+        "//parser/mysql",
+        "//server/internal/util",
+        "@com_github_klauspost_compress//zstd",
+        "@com_github_stretchr_testify//require",
     ],
 )
